idiomatic go (pythonic for go wold) - import organize - const group - named return - is software not package, change name ignore.go to main.go (default main package) - review TODO pls
